refactor: コード生成部分を別ファイルに切り出し
