+++ src/components/projects/ButtonSet.js
@@ -8,6 +8,7 @@ import HammerIcon from "./HammerIcon";
 import Button from "@material-ui/core/Button";
 import { makeStyles } from "@material-ui/core/styles";
 import Modal from "@material-ui/core/Modal";
+import { createNextChord } from "./ChordModel.js";
 
 import notes from "../notes/Notes.js";
 
@@ -201,104 +202,8 @@ const ButtonSet = ({ btnNum, handleChords, setChordNums, randomChord }) => {
         return setCurrentChord(chord.name);
       }
     });
-    const nextChordNums = [];
-    nextChordNums.push(rootNum);
-    switch (chordType) {
-      case "maj":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        break;
-      case "min":
-        nextChordNums.push(rootNum + 3);
-        nextChordNums.push(rootNum + 7);
-        break;
-      case "7":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 10);
-        break;
-      case "M7":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 11);
-        break;
-      case "m7":
-        nextChordNums.push(rootNum + 3);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 10);
-        break;
-      case "mM7":
-        nextChordNums.push(rootNum + 3);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 11);
-        break;
-      case "dim":
-        nextChordNums.push(rootNum + 3);
-        nextChordNums.push(rootNum + 6);
-        nextChordNums.push(rootNum + 9);
-        break;
-      case "sus4":
-        nextChordNums.push(rootNum + 5);
-        nextChordNums.push(rootNum + 7);
-        break;
-      case "7sus4":
-        nextChordNums.push(rootNum + 5);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 10);
-        break;
-      case "aug":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 8);
-        break;
-      case "m7(♭5)":
-        nextChordNums.push(rootNum + 3);
-        nextChordNums.push(rootNum + 6);
-        nextChordNums.push(rootNum + 10);
-        break;
-      case "6":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 9);
-        break;
-      case "add9":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 14);
-        break;
-      case "9":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 10);
-        nextChordNums.push(rootNum + 14);
-        break;
-      case "7(♭9)":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 10);
-        nextChordNums.push(rootNum + 13);
-        break;
-      case "7(♯9)":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 10);
-        nextChordNums.push(rootNum + 15);
-        break;
-      case "m(♭5)":
-        nextChordNums.push(rootNum + 3);
-        nextChordNums.push(rootNum + 6);
-        break;
-      case "(♯5)":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 6);
-        break;
-      case "M7(♯5)":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 6);
-        nextChordNums.push(rootNum + 11);
-        break;
-      default:
-        break;
-    }
+
+    const nextChordNums = createNextChord(rootNum, chordType);
 
     // CreateProjectコンポーネントのstateを更新
     handleChords(
@@ -317,114 +222,17 @@ const ButtonSet = ({ btnNum, handleChords, setChordNums, randomChord }) => {
 
   // propsで渡されたrandomChordを使ってここのボタン情報を更新する
   useEffect(() => {
-    shuffleRootRadio(randomChord[0]);
-    shuffleChordRadio(randomChord[1]);
-
     const rootNum = randomChord[0];
     const chordType = randomChord[1];
     const rootName = notesStrings.find((v) => v.value === rootNum).name;
 
+    shuffleRootRadio(rootNum);
+    shuffleChordRadio(chordType);
+
     setCurrentNote(rootName);
     setCurrentChord(chordType);
 
-    const nextChordNums = [];
-    nextChordNums.push(rootNum);
-    switch (chordType) {
-      case "maj":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        break;
-      case "min":
-        nextChordNums.push(rootNum + 3);
-        nextChordNums.push(rootNum + 7);
-        break;
-      case "7":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 10);
-        break;
-      case "M7":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 11);
-        break;
-      case "m7":
-        nextChordNums.push(rootNum + 3);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 10);
-        break;
-      case "mM7":
-        nextChordNums.push(rootNum + 3);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 11);
-        break;
-      case "dim":
-        nextChordNums.push(rootNum + 3);
-        nextChordNums.push(rootNum + 6);
-        nextChordNums.push(rootNum + 9);
-        break;
-      case "sus4":
-        nextChordNums.push(rootNum + 5);
-        nextChordNums.push(rootNum + 7);
-        break;
-      case "7sus4":
-        nextChordNums.push(rootNum + 5);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 10);
-        break;
-      case "aug":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 8);
-        break;
-      case "m7(♭5)":
-        nextChordNums.push(rootNum + 3);
-        nextChordNums.push(rootNum + 6);
-        nextChordNums.push(rootNum + 10);
-        break;
-      case "6":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 9);
-        break;
-      case "add9":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 14);
-        break;
-      case "9":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 10);
-        nextChordNums.push(rootNum + 14);
-        break;
-      case "7(♭9)":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 10);
-        nextChordNums.push(rootNum + 13);
-        break;
-      case "7(♯9)":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 7);
-        nextChordNums.push(rootNum + 10);
-        nextChordNums.push(rootNum + 15);
-        break;
-      case "m(♭5)":
-        nextChordNums.push(rootNum + 3);
-        nextChordNums.push(rootNum + 6);
-        break;
-      case "(♯5)":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 6);
-        break;
-      case "M7(♯5)":
-        nextChordNums.push(rootNum + 4);
-        nextChordNums.push(rootNum + 6);
-        nextChordNums.push(rootNum + 11);
-        break;
-      default:
-        break;
-    }
+    const nextChordNums = createNextChord(rootNum, chordType);
 
     // CreateProjectコンポーネントのstateを更新
     handleChords(
